requirements.txt
BeautifulSoup4==4.12.2
google==3.0.0
googlesearch-python==1.2.3
lxml==4.9.3
requests==2.31.0
ttkthemes==3.2.2
pandas==2.0.3
